requirements.txt
# requirements.txt
todoist-api-python==2.1.3
boto3==1.26.137
google-cloud-pubsub==2.13.11
sendgrid==6.9.7
pytz==2023.3
google-cloud-functions==1.8.0
